requirements.txt
azure-common==1.1.28
azure-core==1.27.1
azure-identity==1.13.0
azure-mgmt-apimanagement==4.0.0
azure-mgmt-cdn==12.0.0
azure-mgmt-containerinstance==10.1.0
azure-mgmt-containerregistry==10.1.0
azure-mgmt-core==1.4.0
azure-mgmt-dns==8.1.0
azure-mgmt-frontdoor==1.1.0
azure-mgmt-hdinsight==9.0.0
azure-mgmt-network==23.1.0
azure-mgmt-rdbms==10.1.0
azure-mgmt-redis==14.1.0
azure-mgmt-resource==23.0.1
azure-mgmt-search==9.0.0
azure-mgmt-storage==21.0.0
azure-mgmt-trafficmanager==1.0.0
azure-mgmt-web==7.1.0
boto==2.49.0
boto3==1.26.153
botocore==1.29.153
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
docutils==0.20.1
google-api-core==2.11.1
google-auth==2.20.0
google-cloud==0.34.0
google-cloud-compute==1.11.0
google-cloud-core==2.3.2
google-cloud-dns==0.34.1
google-cloud-functions==1.13.0
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-oauth==1.0.1
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpc-google-iam-v1==0.12.6
grpcio==1.54.2
grpcio-status==1.54.2
idna==3.4
isodate==0.6.1
jmespath==1.0.1
msal==1.22.0
msal-extensions==1.0.0
msrest==0.7.1
numpy==1.25.0
oauthlib==3.2.2
pandas==2.0.2
portalocker==2.7.0
proto-plus==1.22.2
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
PyMySQL==1.0.3
pyOpenSSL==23.2.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.1
six==1.16.0
tabulate==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
